working updates to nonstationary noise reduction
