+++ noisereduce/spectralgate/__init__.py
@@ -1,7 +1,6 @@
-from .nonstationary import SpectralGateNonStationary
-from .stationary import SpectralGateStationary
 try:
     import torch
+
     TORCH_AVAILABLE = True
 except ImportError:
     TORCH_AVAILABLE = False
