--- noisereduce/spectralgate/nonstationary.py
@@ -1,115 +1,0 @@
-from noisereduce.spectralgate.base import SpectralGate
-import numpy as np
-from scipy.signal import filtfilt, fftconvolve, stft, istft
-from .utils import sigmoid
-
-
-class SpectralGateNonStationary(SpectralGate):
-    def __init__(
-            self,
-            y,
-            sr,
-            chunk_size,
-            padding,
-            n_fft,
-            win_length,
-            hop_length,
-            time_constant_s,
-            freq_mask_smooth_hz,
-            time_mask_smooth_ms,
-            thresh_n_mult_nonstationary,
-            sigmoid_slope_nonstationary,
-            tmp_folder,
-            prop_decrease,
-            use_tqdm,
-            n_jobs,
-    ):
-        self._thresh_n_mult_nonstationary = thresh_n_mult_nonstationary
-        self._sigmoid_slope_nonstationary = sigmoid_slope_nonstationary
-
-        super().__init__(
-            y=y,
-            sr=sr,
-            chunk_size=chunk_size,
-            padding=padding,
-            n_fft=n_fft,
-            win_length=win_length,
-            hop_length=hop_length,
-            time_constant_s=time_constant_s,
-            freq_mask_smooth_hz=freq_mask_smooth_hz,
-            time_mask_smooth_ms=time_mask_smooth_ms,
-            tmp_folder=tmp_folder,
-            prop_decrease=prop_decrease,
-            use_tqdm=use_tqdm,
-            n_jobs=n_jobs,
-        )
-
-    def spectral_gating_nonstationary(self, chunk):
-        """non-stationary version of spectral gating"""
-        denoised_channels = np.zeros(chunk.shape, chunk.dtype)
-        for ci, channel in enumerate(chunk):
-            _, _, sig_stft = stft(
-                channel,
-                nfft=self._n_fft,
-                noverlap=self._win_length - self._hop_length,
-                nperseg=self._win_length,
-                padded=False
-            )
-            # get abs of signal stft
-            abs_sig_stft = np.abs(sig_stft)
-
-            # get the smoothed mean of the signal
-            sig_stft_smooth = get_time_smoothed_representation(
-                abs_sig_stft,
-                self.sr,
-                self._hop_length,
-                time_constant_s=self._time_constant_s,
-            )
-
-            # get the number of X above the mean the signal is
-            sig_mult_above_thresh = (abs_sig_stft - sig_stft_smooth) / sig_stft_smooth
-            # mask based on sigmoid
-            sig_mask = sigmoid(
-                sig_mult_above_thresh,
-                -self._thresh_n_mult_nonstationary,
-                self._sigmoid_slope_nonstationary,
-            )
-
-            if self.smooth_mask:
-                # convolve the mask with a smoothing filter
-                sig_mask = fftconvolve(sig_mask, self._smoothing_filter, mode="same")
-
-            sig_mask = sig_mask * self._prop_decrease + np.ones(np.shape(sig_mask)) * (
-                    1.0 - self._prop_decrease
-            )
-
-            # multiply signal with mask
-            sig_stft_denoised = sig_stft * sig_mask
-
-            # invert/recover the signal
-            _, denoised_signal = istft(
-                sig_stft_denoised,
-                nfft=self._n_fft,
-                noverlap=self._win_length - self._hop_length,
-                nperseg=self._win_length
-            )
-            denoised_channels[ci, : len(denoised_signal)] = denoised_signal
-        return denoised_channels
-
-    def _do_filter(self, chunk):
-        """Do the actual filtering"""
-        chunk_filtered = self.spectral_gating_nonstationary(chunk)
-
-        return chunk_filtered
-
-
-def get_time_smoothed_representation(
-        spectral, samplerate, hop_length, time_constant_s=0.001
-):
-    t_frames = time_constant_s * samplerate / float(hop_length)
-    # By default, this solves the equation for b:
-    #   b**2  + (1 - b) / t_frames  - 2 = 0
-    # which approximates the full-width half-max of the
-    # squared frequency response of the IIR low-pass filt
-    b = (np.sqrt(1 + 4 * t_frames ** 2) - 1) / (2 * t_frames ** 2)
-    return filtfilt([b], [1, b - 1], spectral, axis=-1, padtype=None)
