+++ noisereduce/spectralgate/streamed_torch_gate.py
@@ -1,47 +1,49 @@
 import torch
-from noisereduce.spectralgate.base import SpectralGate
+from noisereduce.spectralgate.spectralgate import SpectralGate
 from noisereduce.torchgate import TorchGate as TG
 import numpy as np
 
 
 class StreamedTorchGate(SpectralGate):
-    '''
+    """
     Run interface with noisereduce.
-    '''
+    """
 
     def __init__(
-            self,
-            y,
-            sr,
-            stationary=False,
-            y_noise=None,
-            prop_decrease=1.0,
-            time_constant_s=2.0,
-            freq_mask_smooth_hz=500,
-            time_mask_smooth_ms=50,
-            thresh_n_mult_nonstationary=2,
-            sigmoid_slope_nonstationary=10,
-            n_std_thresh_stationary=1.5,
-            tmp_folder=None,
-            chunk_size=600000,
-            padding=30000,
-            n_fft=1024,
-            win_length=None,
-            hop_length=None,
-            clip_noise_stationary=True,
-            use_tqdm=False,
-            n_jobs=1,
-            device="cuda",
+        self,
+        y,
+        sr,
+        stationary=False,
+        y_noise=None,
+        prop_decrease=1.0,
+        noise_window_size_nonstationary_ms=None,
+        freq_mask_smooth_hz=500,
+        time_mask_smooth_ms=50,
+        n_std_thresh=1.5,
+        tmp_folder=None,
+        chunk_size=600000,
+        padding=30000,
+        n_fft=1024,
+        win_length=None,
+        hop_length=None,
+        clip_noise_stationary=True,
+        use_tqdm=False,
+        n_jobs=1,
+        device="cuda",
     ):
         super().__init__(
             y=y,
             sr=sr,
+            stationary=stationary,
+            y_noise=y_noise,
             chunk_size=chunk_size,
             padding=padding,
             n_fft=n_fft,
+            n_std_thresh=n_std_thresh,
+            noise_window_size_nonstationary_ms=noise_window_size_nonstationary_ms,
             win_length=win_length,
             hop_length=hop_length,
-            time_constant_s=time_constant_s,
+            clip_noise_stationary=clip_noise_stationary,
             freq_mask_smooth_hz=freq_mask_smooth_hz,
             time_mask_smooth_ms=time_mask_smooth_ms,
             tmp_folder=tmp_folder,
@@ -50,7 +52,7 @@ class StreamedTorchGate(SpectralGate):
             n_jobs=n_jobs,
         )
 
-        self.device = torch.device(device if torch.cuda.is_available() else 'cpu')
+        self.device = torch.device(device if torch.cuda.is_available() else "cpu")
 
         # noise convert to torch if needed
         if y_noise is not None:
@@ -66,10 +68,7 @@ class StreamedTorchGate(SpectralGate):
         self.tg = TG(
             sr=sr,
             nonstationary=not stationary,
-            n_std_thresh_stationary=n_std_thresh_stationary,
-            n_thresh_nonstationary=thresh_n_mult_nonstationary,
-            temp_coeff_nonstationary=1 / sigmoid_slope_nonstationary,
-            n_movemean_nonstationary=int(time_constant_s / self._hop_length * sr),
+            n_std_thresh=n_std_thresh,
             prop_decrease=prop_decrease,
             n_fft=self._n_fft,
             win_length=self._win_length,
